requirements.txt
fastapi
psycopg2-binary
python-dotenv
openai
pydantic
uvicorn

